fix: vo로 input validation 분리
