+++ src/main/java/racing/domain/RacingGame.java
@@ -1,7 +1,5 @@
 package racing.domain;
 
-import racing.Constant;
-import racing.dto.GameVO;
 import racing.utils.RandomUtil;
 import racing.view.ResultView;
 
@@ -14,16 +12,6 @@ public class RacingGame {
     public RacingGame(Cars cars, int round) {
         this.cars = cars;
         this.round = round;
-    }
-
-    public void validateInput(GameVO vo) {
-        if (vo.getCarNo() <= 0) {
-            throw new IllegalArgumentException(Constant.CAR_NUMBER_VALIDATION_ERROR);
-        }
-
-        if (vo.getRoundNo() <= 0) {
-            throw new IllegalArgumentException(Constant.ROUND_NUMBER_VALIDATION_ERROR);
-        }
     }
 
     public void play() {
